refactor(rust): `spending.rs` -> `spends.rs`
